BUG: ensure that casting to/from structured is properly checked.
Allow unsafe casting from a simple data type to a structured one with
multiple fields, but only from structured data type with a single
field (checked recursively) to a simple data type.

For now, continue to allow any structured to structured with
casting="unsafe", as the current can_cast implementation does not
match that of "astype".