MAINT: Test names against None for consistency
In these instances the behavior isn't changed, since the for loop below acts like an if.
However, in general this is an antipattern that crashes on 0-field structured types, and is warned against in the docs.
If we remove instances of the antipattern, it will hopefully not reappear via copy-paste code.