MAINT: Only allow using Cython module when cythonizing.
The command line version of `cython` may point to a different
installation of Cython than that installed in the Python running the
cythonize script. Because the Cython version can be critical, requiring
that the cython used comes from a known place makes it less likely that
the wrong version will be used. This treats Cython as a build dependency
rather than a free standing compiler.