ENH,DEP: Allow multiple axes in expand_dims (#14051)
This PR allows the axis keyword in expand_dims to be a tuple of ints. Previously, axis could only be an int.

This issue was previously discussed in gh-12290 and the changes are based on gh-12290 (comment).

This PR also removes the deprecation added in v1.13 (2017-05-17), where previously axis could be outside of the range (-a.ndim - 1) <= axis <= a.ndim. Such an axis value will now raise an AxisError. Please let me know if it's too soon to remove this deprecation (I could not find any dev docs stating the length of the numpy deprecation cycle).

Closes gh-12290.