BUG: `np.resize` negative shape and subclasses edge case fixes
When a negative entry is passed into `np.resize` shape, an error will
now be raised reliably. Previously the `-1` semantics of reshape allowed
edge cases to pass and return incorrect results.
In the corner case of an empty result shape or empty input shape, the
result type will now be preserved.