BUG: distutils: fix msvc+gfortran openblas handling corner case
Ensure the openblas MSVC+gfortran temporary library names are unique
for the different openblas_* system_info classes.

If multiple openblas libraries (e.g. both a 32-bit and a 64-bit one) are
used in the same project, when compiling in the msvc+gfortran mode, this
previously resulted to only the last one being used.