MAINT: simd: Avoid signed comparison warning
We do not need to cast the dimensions (`n`) to ``npy_uintp`` when
comparing it to the stride length (of type ``npy_intp``). The cast is
kept from similar checks where the cast is required

    ``(npy_uintp)n < (VECTOR_SIZE_BYTES / sizeof(@type@))``

In this case `strides` is precalcuated as

    ``VECTOR_SIZE_BYTES / (npy_intp)sizeof(@type@))``

therefore `n` is already of the correct type when the check is
performed.