BUG: Do not dispatch to base classes polymorphically
The previously code was equivalent to something like
```python
class SomeScalar(base1, base2):
    def __new__(cls, *args):
        return cls.__bases__[1].__new__(*args)
```
This is nonsense and can easily cause recursion, for instance if `cls` is a superclass of SomeScalar.
The correct way to spell this is to pin the base class call statically.
```python
class SomeScalar(base1, base2):
    def __new__(cls, *args):
        return SomeScalar.__bases__[1].__new__(*args)
```

Fixes #15395