DEP: Schedule PyArray_GetArrayParamsFromObject for removal
The function has no known usage outside of NumPy. It is fairly
complex to use due to how it behaves with scalars, and, at least
with a given "flexible" dtype would require additional access
to the private PyArray_AdaptFlexibleDType function.

Removing the function from public API will simplify refactor work
of the dtype and dimension recovery.