DEP: Deprecate PyUFunc_SetUsesArraysAsData
This function is currently undocumented. It is not used in NumPy
and seems useless outside of NumPy. This is due to the fact that
user datatypes are not actually capable of leveraging this function!
They are stored in a different way, so that their loops cannot be
identified by an index.