API: Create Preliminary DTypeMeta class and np.dtype subclasses
This makes all dtypes being subclasses of a:

class DType(np.dtype):
    is_abstract = False
    is_flexible = False  % True for void, strings, and datetimes
    singleton = prototype_dtype_instance
    # other information defined on the singleton currently

subclass of np.dtype.

Right now this happens by replacing the Type field of of the
singleton instance. This is valid for types/objects not created
in python (not HeapTypes). It is necessary to have this functionality
for continuing support of user defined legacy dtypes.