DOC,BLD: Limit timeit iterations in random docs.
Limiting the number of loops performed in the %timeit evaluation
results in a ~20% speedup in the time it takes to build the
documentation.

The results of the "benchmarking" are less accurate from the limited
number of evaluations, but the relationship between the timing
between the legacy and Generator sampling methods is preserved.