BENCH: Default to building HEAD instead of master
asv executes the benchmarks of the current branch building the code
specified in "branches". Previously this was "master" instead of "HEAD"
(the code currently worked on). This PR changes "master" -> "HEAD" as it
seems more likely for a user to be generating benchmarks for the code
they are currently working on.