BUG,MAINT: Simplify reduction result creation using reduce axis
The new reduce axis marker for nditer/NpyIter allows to move the
allocation logic necessary to allocate the reduction result
to the nditer, thus simplifying the logic here.

This leads to some other related cleanups of the result
initialization.
The subok flag was removed, since it was always set to 0
and the iterator rejects subclasses.