MAINT: Simplify ufunc.outer and raise error on bad reshape
This simplifies the reshaping logic to use the NewShape function
(this was previously impossible due to the split of umath and
multiarray).
It adds new errors if reshapes are bad, which currently is only
possible for bad subclasses. However, `np.matrix` is the main bad
subclass here, so this is in preparation of deprecating it.