DEP: Deprecate ufunc.outer with matrix inputs
Previously we converted matrices to arrays for the user, since
`np.outer` does not make sense for matrix. This deprecates that
usage (which in almost all cases will result in an error).

The main reason for this deprecation is to remove all direct
references of matrix from the NumPy codebase, so that it is
easier to move it out of NumPy eventually.