Replace AdaptFlexibleDType with object and delete some datetime thing related
This fails two tests (but behaviour is better now). The string->datetime
unit discovery for string arrays is hard-coded now, and IMO should be
deprecated at some point, if necessary we can add a specific function
for it maybe?

Unless we really want to make this pattern a first class citizen (and
not special case object). But I am against that until someone finds a
better use-case...