MAINT: User array alignment to speed up buffer export
Calculating alignments can be slow, but unless we have a structured
dtype, we can simply use the arrays alignment.
This code can probably re-use more of the array alignment code
(or maybe alignment checks moved into the dtype) even in the
structured case.  The code using the % operator is very slow
compared to the existing power-of-two-assuming alignment code.