BUG: Initialize stop-reading in array_from_text
This should probably be hit with count=0, which is why added
a test which takes that path.  However, I was not actually
able to trigger any issue (not even a warning in valgrind).
I assume that since this is on the stack, its just hard to
trigger.
This came up in another CI-run. I suppose gcc as well as clang seem
to notice that it can be used uninitialized (which is true).